+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @reservoir0x/relay-kit-hooks
 
+## 3.0.8
+
+### Patch Changes
+
+- Updated dependencies [f851b3c]
+  - @relayprotocol/relay-sdk@5.1.1
+
+## 3.0.7
+
+### Patch Changes
+
+- 448ec7c: Refactor deposit address transaction tracking to use requests API
+
 ## 3.0.6
 
 ### Patch Changes